Centralize StyleX functional lowering
Amp-Thread-ID: https://ampcode.com/threads/T-01a062c8-25a2-77e6-a3e8-2fb6c000398d
