+++ compiler-backend/functional/src/convert.rs
@@ -4,6 +4,7 @@ mod application;
 mod declaration;
 mod evidence;
 mod expression;
+mod stylex;
 
 use std::sync::Arc;
 
@@ -18,12 +19,12 @@ use smol_str::{SmolStr, format_smolstr};
 use thiserror::Error;
 
 use crate::error::{ModuleError, ModuleResult, UnsupportedState};
-use crate::optimize::{expression_children, inline_simple_bindings};
+use crate::optimize::inline_simple_bindings;
 use crate::tree::{
     Declaration, DeclarationKind, Expression, ExpressionId, ExpressionKind, Field, FieldIdentity,
     GeneratedGlobalId, Global, GlobalId, IndirectModuleExports, InstanceIdentity, LocalId, Module,
     ModuleDependency, ModuleSurface, Parameter, Pattern, PatternId, PatternKind, RecursiveGroupId,
-    Storage, StyleXIntrinsic, SuperclassIdentity,
+    Storage, SuperclassIdentity,
 };
 
 use self::declaration::{derive_declaration, instance_declaration, term_declaration};
@@ -605,53 +606,6 @@ where
         Ok(Global { id: GlobalId::Term(file_id, term_id), item_name })
     }
 
-    fn stylex_intrinsic_identity(
-        &self,
-        file_id: FileId,
-        term_id: TermItemId,
-    ) -> QueryResult<Option<StyleXIntrinsic>> {
-        if self.queries.module_file("Alexandrite.StyleX") != Some(file_id) {
-            return Ok(None);
-        }
-        let indexed = self.indexed_module(file_id)?;
-        let intrinsic = match indexed.items[term_id].name.as_deref() {
-            Some("create") => Some(StyleXIntrinsic::Create),
-            Some("props") => Some(StyleXIntrinsic::Props),
-            Some("recordProps") => Some(StyleXIntrinsic::RecordProps),
-            Some("conditional") => Some(StyleXIntrinsic::Conditional),
-            Some("keyframes") => Some(StyleXIntrinsic::Keyframes),
-            _ => None,
-        };
-        Ok(intrinsic)
-    }
-
-    fn validate_stylex_uses(&self, declarations: &[Declaration]) -> ConversionResult<()> {
-        let pending = declarations.iter().filter_map(|declaration| match declaration.kind {
-            DeclarationKind::Value(expression) => Some((expression, declaration.global.id)),
-            DeclarationKind::Constructor { .. } | DeclarationKind::Foreign => None,
-        });
-        let mut pending = pending.collect_vec();
-        let mut visited = FxHashSet::default();
-        while let Some((expression, declaration)) = pending.pop() {
-            if !visited.insert(expression) {
-                continue;
-            }
-            if let ExpressionKind::Global { global } = &self.storage[expression].kind
-                && let GlobalId::Term(file_id, term_id) = global.id
-                && let Some(intrinsic) = self.stylex_intrinsic_identity(file_id, term_id)?
-            {
-                let state = UnsupportedState::InvalidStyleXUse {
-                    function: intrinsic.name().to_owned(),
-                    declaration,
-                };
-                return Err(self.unsupported(state));
-            }
-            let children = expression_children(&self.storage[expression].kind);
-            pending.extend(children.into_iter().map(|expression| (expression, declaration)));
-        }
-        Ok(())
-    }
-
     fn constructor_is_newtype(&self, file_id: FileId, term_id: TermItemId) -> QueryResult<bool> {
         let indexed = self.indexed_module(file_id)?;
         let Some(type_id) = indexed.constructor_type(term_id) else {
@@ -708,29 +662,6 @@ where
         let module_name = self.source_module_name(file_id)?;
         self.dependencies.insert(file_id, Dependency { module_name, indexed });
         Ok(())
-    }
-
-    fn module_is_virtual(&self, file_id: FileId) -> bool {
-        self.queries.module_file("Alexandrite.StyleX") == Some(file_id)
-    }
-
-    fn validate_runtime_reference(
-        &self,
-        file_id: FileId,
-        term_id: TermItemId,
-    ) -> ConversionResult<()> {
-        if !self.module_is_virtual(file_id) {
-            return Ok(());
-        }
-        let module_name = self.source_module_name(file_id)?.to_string();
-        let indexed = self.indexed_module(file_id)?;
-        let item_name = indexed.items[term_id]
-            .name
-            .clone()
-            .unwrap_or_else(|| self.term_fallback(term_id))
-            .to_string();
-        let state = UnsupportedState::VirtualModuleRuntimeReference { module_name, item_name };
-        Err(self.unsupported(state))
     }
 
     fn instance_name(&self, identity: InstanceIdentity) -> QueryResult<SmolStr> {
